Prepare build script for use in ci workflow
This commit adds an additional check to the build script so the llvm
repo is only cloned if the directory for it does not yet exist.
Otherwise only a pull to fetch the latest version is performed.
The "no-llvm" option is preserved for compatibility and in case any
updates to the llvm build need to be skipped.

Additional compiler and linker specific options to the cinnamon
configuration step where removed as they should be handled via
environment variables on a machine by machine basis. E.g.:
export CC=clang
export CXX=clang++
export CMAKE_GENERATOR=Ninja
export LDFLAGS="-fuse-ld=mold"

Additional flags for the cmake configuration of cinnamon (like UPMEM_DIR)
can be passed via the CINNAMON_CMAKE_OPTIONS environment variable.

The specification of the UPMEM_DIR should probably be handled
in a more flexible way in the future, e.g. via a find_package in cmake.