Add build and test ci workflow.
This commit adds a github actions workflow for building and testing
cinnamon. It is triggered automatically on every push or manually via
workflow dispatch.

The llvm build is cached between runs to reduce unnecessary compilation
work and build times. The full uncached build, which will have to be
performed one time to build the cache, takes a few hours.

The cinnamon build is done completely on every run. A build failiure will
result in a workflow fail.

For testing check-cinm-mlir will be executed. Should any test registered
to this executable fail the whole workflow will also fail.

This ci workflow should help find problematic commit and pull requests
which would result in a broken build or regression.