Integrate tilefirst as depenedncy
Adopt the contents of the old upmemx dialect.
The one problem is that I can't find a way to
mark the imported tilefirst headers as system.
One solution for this would be to stop using the
TileFirstConfig.cmake generator, and just add a
global dependency on Tilefirst. Maybe build an
aggregate library with all TF libraries. We would
lose fine-grained dependency spec though.