Do transfer elimination manually
This is hard because in order to reuse
an mram buffer that is declared inside
a `tile[r*d] hwparallel factor<R*D>`, you
need to pull it out and tile it because of
the rules on parallel loop schedulers and
interference. That means you need a buffer
`<2**20, f32, mram(r,d)>` which is _distributed_,
ie, it is only used to be tiled with those
tile loops, and therefore only the tile actually
exists in memspace `mram(r,d)`. But you need to
pull it out if you want to track its values across
several of these tile loops and eliminate the redundant transfers.

The remaining paint point will be to generate an
upmem DPU program which has one of three states.
We cannot reload a program on the DPU set without
clearing all the global data, therefore we would
be deleting a reusable buffer. So we will need the
upmem generation to support the generation of an
automaton.